Reorganize account logic into services and correctly arrange repositories. Add missing DTOs.

+++ Vijuge/Vijuge.Data/Models/DTOs/PlayerDTO.cs
@@ -9,6 +9,6 @@
 
         public bool LoggedIn { get; set; }
         public int Rank { get; set; }
-        public long PointsCollected { get; set; }
+        public double PointsCollected { get; set; }  
     }
 }
